pyproject.toml
[project]
name = "yt-rapport"
version = "0.1.0"
description = "CLI pour extraire et analyser le contenu de vidéos YouTube"
requires-python = ">=3.12"
dependencies = [
    "typer>=0.12",
    "rich>=13",
    "youtube-transcript-api>=0.6",
    "openai>=1.0",
    "python-dotenv>=1.0",
    "httpx>=0.27",
]

[project.scripts]
yt-rapport = "yt_rapport.main:app"

[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[tool.hatch.build.targets.wheel]
packages = ["src/yt_rapport"]
